+++ Src/Commands/AddDataCommand.cs
@@ -4,8 +4,6 @@
 
 using LiteDB;
 using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Management.Automation;
 using System.Reflection;
 
@@ -32,17 +30,11 @@ namespace Ldbc.Commands
 		[Parameter]
 		public SwitchParameter Result { get; set; }
 
-		[Parameter]
-		public SwitchParameter Bulk { get; set; }
-		List<object> _bulk;
-
-		BsonAutoId _autoId; //rk temp
+		System.Func<object, BsonDocument> _convert;
 
 		protected override void BeginProcessing()
 		{
-			_autoId = Collection.AutoId();
-			if (Bulk)
-				_bulk = new List<object>();
+			_convert = Actor.ToBsonDocumentNoDefaultId(Collection.AutoId());
 		}
 
 		protected override void ProcessRecord()
@@ -50,16 +42,9 @@ namespace Ldbc.Commands
 			if (InputObject == null)
 				throw new PSArgumentException(Res.InputDocNull);
 
-			if (Bulk)
-			{
-				_bulk.Add(InputObject);
-				return;
-			}
-
 			try
 			{
-				var document = Actor.ToBsonDocument(InputObject);
-				Actor.RemoveDefaultId(document, _autoId);
+				var document = _convert(InputObject);
 				var id = Collection.Insert(document);
 				if (Result)
 					WriteObject(Actor.ToObject(id));
@@ -72,17 +57,6 @@ namespace Ldbc.Commands
 			{
 				WriteException(exn, InputObject);
 			}
-		}
-
-		protected override void EndProcessing()
-		{
-			if (!Bulk)
-				return;
-
-			var convert = Actor.ToBsonDocumentNoDefaultId(_autoId);
-			var count = Collection.Insert(_bulk.Select(convert));
-			if (Result)
-				WriteObject(count);
 		}
 	}
 }